backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0

# OpenAI Integration
openai==1.12.0

# Database & ORM
sqlalchemy==2.0.27
asyncpg==0.29.0
supabase==2.4.0

# Geospatial
geopy==2.4.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0  # For testing async HTTP calls

# Development
black==24.2.0
ruff==0.2.2
mypy==1.8.0
